requirements.txt
struct
numpy
matplotlib.pyplot
math
os
time